relocate framework and loader in to improve workflow
while iterating on the framework, and the loader and a sample
it can be tedious to copy from a few different spots. This should
improve that workflow, also likely makes the project easier to
work with because the need to copy DLLs from a different repo
are gone. Now its just setting up Thunderkit for Aloft.